Refactor contrast outcome
