+++ packages/alfa-rules/src/sia-r66/rule.ts
@@ -12,7 +12,6 @@ import { Iterable } from "@siteimprove/alfa-iterable";
 import { Option, None } from "@siteimprove/alfa-option";
 import { Predicate } from "@siteimprove/alfa-predicate";
 import { Refinement } from "@siteimprove/alfa-refinement";
-import { Ok, Err } from "@siteimprove/alfa-result";
 import { Style } from "@siteimprove/alfa-style";
 import { Criterion } from "@siteimprove/alfa-wcag";
 import { Page } from "@siteimprove/alfa-web";
@@ -23,6 +22,8 @@ import { hasAttribute } from "../common/predicate/has-attribute";
 import { hasRole } from "../common/predicate/has-role";
 import { hasValue } from "../common/predicate/has-value";
 import { isPerceivable } from "../common/predicate/is-perceivable";
+
+import { Outcomes } from "../common/outcome/contrast";
 
 import { Question } from "../common/question";
 
@@ -134,36 +135,6 @@ export default Rule.Atomic.of<Page, Text, Question>({
     };
   },
 });
-
-export namespace Outcomes {
-  export const HasSufficientContrast = (
-    highest: number,
-    threshold: number,
-    pairings: Array<Contrast.Pairing>
-  ) =>
-    Ok.of(
-      Contrast.of(
-        `The highest possible contrast of the text is 1:${highest} which is
-        above the required contrast of 1:${threshold}`,
-        threshold,
-        pairings
-      )
-    );
-
-  export const HasInsufficientContrast = (
-    highest: number,
-    threshold: number,
-    pairings: Array<Contrast.Pairing>
-  ) =>
-    Err.of(
-      Contrast.of(
-        `The highest possible contrast of the text is 1:${highest} which is
-        below the required contrast of 1:${threshold}`,
-        threshold,
-        pairings
-      )
-    );
-}
 
 /**
  * @see https://w3c.github.io/wcag/guidelines/#dfn-large-scale
