+++ packages/alfa-rules/src/sia-r69/rule.ts
@@ -24,6 +24,8 @@ import { hasRole } from "../common/predicate/has-role";
 import { hasValue } from "../common/predicate/has-value";
 import { isPerceivable } from "../common/predicate/is-perceivable";
 
+import { Outcomes } from "../common/outcome/contrast";
+
 import { Question } from "../common/question";
 
 import { Contrast } from "../common/diagnostic/contrast";
@@ -134,36 +136,6 @@ export default Rule.Atomic.of<Page, Text, Question>({
     };
   },
 });
-
-export namespace Outcomes {
-  export const HasSufficientContrast = (
-    highest: number,
-    threshold: number,
-    pairings: Array<Contrast.Pairing>
-  ) =>
-    Ok.of(
-      Contrast.of(
-        `The highest possible contrast of the text is 1:${highest} which is
-        above the required contrast of 1:${threshold}`,
-        threshold,
-        pairings
-      )
-    );
-
-  export const HasInsufficientContrast = (
-    highest: number,
-    threshold: number,
-    pairings: Array<Contrast.Pairing>
-  ) =>
-    Err.of(
-      Contrast.of(
-        `The highest possible contrast of the text is 1:${highest} which is
-        below the required contrast of 1:${threshold}`,
-        threshold,
-        pairings
-      )
-    );
-}
 
 /**
  * @see https://w3c.github.io/wcag/guidelines/#dfn-large-scale
