+++ packages/alfa-rules/tsconfig.json
@@ -10,6 +10,7 @@
     "src/common/expectation/outcome-to-trilean.ts",
     "src/common/expectation/video-description-track-accurate.ts",
     "src/common/group.ts",
+    "src/common/outcome/contrast.ts",
     "src/common/outcome/text-spacing.ts",
     "src/common/predicate/has-accessible-name.ts",
     "src/common/predicate/has-attribute.ts",
